Consider more points for evaluation, it still sucks

+++ src/main/scala/package.scala
@@ -183,7 +183,7 @@ package object zahak {
         case `BN` | `WN` => 300
         case `BR` | `WR` => 500
         case `BQ` | `WQ` => 900
-        case `BK` | `WK` => Double.MaxValue
+        case `BK` | `WK` => Int.MaxValue
       }
   }
